requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.8.0
pdf2image>=1.16.3
pandas>=2.0.0
openpyxl>=3.1.0
ollama>=0.1.7
numpy>=1.24.0
scikit-image>=0.21.0  # For additional image processing
matplotlib>=3.7.0  # For image visualization if needed
requests>=2.31.0  # For API calls